+++ image_dds/src/error.rs
@@ -5,7 +5,7 @@ use crate::DdsFormatInfo;
 use crate::ImageFormat;
 
 /// Errors that can occur while creating a decoded image.
-#[derive(Debug, Error)]
+#[derive(Debug, Error, PartialEq)]
 pub enum CreateImageError {
     #[error("data length {data_length} is not valid for a {width}x{height} image")]
     InvalidSurfaceDimensions {
@@ -22,7 +22,7 @@ pub enum CreateImageError {
 }
 
 /// Errors that can occur while encoding or decoding a surface.
-#[derive(Debug, Error)]
+#[derive(Debug, Error, PartialEq)]
 pub enum SurfaceError {
     #[error("surface dimensions {width} x {height} x {depth} contain no pixels")]
     ZeroSizedSurface { width: u32, height: u32, depth: u32 },